+++ Assets/BulkTools/BulkToolsManager/Editor/AssemblyDefTools/ModuleMetaData.cs
@@ -14,7 +14,30 @@ namespace BTools.Management.EditorScripts
         public string editorAssembly = "";
         public List<string> supportedPlatforms = new List<string>();
 
-        public List<string> dependencies = new List<string>();
+        [Tooltip("Dependencies of other modules")]
+        public List<string> internalDependencies = new List<string>();
+
+        [System.Serializable]
+        public struct ExternalPackage 
+        {
+            public string displayName;
+            public string name;
+            public string source;
+        }
+
+        [Tooltip("Dependencies of other packages")]
+        public List<ExternalPackage> externalDependencies = new List<ExternalPackage>();
+
+
+        [System.Serializable]
+        public struct DefineSetting 
+        {
+            public string defineSymbol;
+            [TextArea]
+            public string description;
+        }
+
+        public List<DefineSetting> defineSettings = new List<DefineSetting>();
 
         public string GetModuleRootPath() 
         {